# chore: improve logs displayed when scheduling fails