Fixes # <!-- issue number -->

**Description**
improve logs displayed when scheduling fails
**How was this change tested?**


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.