@tzneal 2023-05-03
New sample error messages:
```
  logger.go:130: 2023-05-02T21:34:15.278-0500	ERROR	scheduling/scheduler.go:181	Could not schedule pod, incompatible with provisioner "raccoonpatch-2255-qybye8gsgg", no instance type satisfied resources {"pods":"1"} and requirements karpenter.sh/capacity-type In [on-demand spot], karpenter.sh/provisioner-name In [raccoonpatch-2255-qybye8gsgg], topology.kubernetes.io/zone In [non-existent-zone], testing.karpenter.sh/test-id In [unspecified] (no instance type met the scheduling requirements or had a required offering)	{"pod": "default/snoutkiwi-2257-3mshp0o3xd"}
```

```
logger.go:130: 2023-05-02T21:34:23.271-0500	ERROR	scheduling/scheduler.go:181	Could not schedule pod, incompatible with provisioner "tigercotton-2893-ipk2xrqnuh", no instance type satisfied resources {"pods":"1"} and requirements topology.kubernetes.io/zone In [test-zone-3], karpenter.sh/capacity-type In [spot], testing.karpenter.sh/test-id In [unspecified], karpenter.sh/provisioner-name In [tigercotton-2893-ipk2xrqnuh] (no instance type has the required offering)	{"pod": "default/howlershard-2897-ap1g0csuzi"}
```

```
logger.go:130: 2023-05-02T21:34:48.181-0500	ERROR	scheduling/scheduler.go:181	Could not schedule pod, incompatible with provisioner "sharkcrocus-4397-c1uqj7foaf", no instance type satisfied resources {"cpu":"14","pods":"1"} and requirements kubernetes.io/arch In [amd64], testing.karpenter.sh/test-id In [unspecified], karpenter.sh/provisioner-name In [sharkcrocus-4397-c1uqj7foaf] (no instance type which had enough resources and the required offering met the scheduling requirements)	{"pod": "default/finchip-4399-3fplmf5fem"}
```