@tzneal 2023-05-05
Negligible difference in scheduling performance:

```
name               old time/op   new time/op   delta
Scheduling1-12       364µs ± 1%    367µs ± 1%  +1.06%  (p=0.032 n=5+5)
Scheduling50-12     18.6ms ± 2%   19.0ms ± 5%    ~     (p=0.151 n=5+5)
Scheduling100-12    37.0ms ± 2%   37.3ms ± 3%    ~     (p=0.690 n=5+5)
Scheduling500-12     181ms ± 4%    180ms ± 5%    ~     (p=1.000 n=5+5)
Scheduling1000-12    540ms ±74%    388ms ±12%    ~     (p=0.841 n=5+5)
Scheduling2000-12    743ms ± 9%    784ms ± 5%    ~     (p=0.151 n=5+5)
Scheduling5000-12    2.01s ± 3%    1.93s ± 5%    ~     (p=0.310 n=5+5)
```