@jackfrancis 2023-05-08
(I see this just merged, but I'll keep asking questions :))

Are these scheduling failures during scale out considerations? Or are these orthogonal to scale out/in and evaluated based upon nodes and their capabilities that are actually running in the cluster at the time?