@tzneal 2023-05-08
> Are these scheduling failures during scale out considerations? Or are these orthogonal to scale out/in and evaluated based upon nodes and their capabilities that are actually running in the cluster at the time?

They occur during both provisioning and consolidation, but they're not logged during consolidation.  The same scheduling code is used for both, consolidation is essentially just:

1) Simulate the effect of one or more nodes being deleted by re-scheduling their pods
2) It the result cheaper?
3) If yes, do that, otherwise try again on a different candidate(s)